[MODUSERBL-169] - Refactoring
